refactor: one useWrap file
