+++ src/hooks/swap/useWrap.ts
@@ -1,5 +1,6 @@
 import { ContractTransaction } from '@ethersproject/contracts'
 import { useWeb3React } from '@web3-react/core'
+import { WRAPPED_NATIVE_CURRENCY } from 'constants/tokens'
 import { useWETHContract } from 'hooks/useContract'
 import { useAtomValue } from 'jotai/utils'
 import { useMemo } from 'react'
@@ -9,7 +10,27 @@ import tryParseCurrencyAmount from 'utils/tryParseCurrencyAmount'
 
 import useCurrencyBalance from '../useCurrencyBalance'
 
-export default function useWrapCallback(
+export function useWrapType(): TransactionType.WRAP | TransactionType.UNWRAP | undefined {
+  const { chainId } = useWeb3React()
+  const { [Field.INPUT]: inputCurrency, [Field.OUTPUT]: outputCurrency } = useAtomValue(swapAtom)
+  return useMemo(() => {
+    if (chainId && inputCurrency && outputCurrency) {
+      if (inputCurrency.isNative && WRAPPED_NATIVE_CURRENCY[chainId]?.equals(outputCurrency)) {
+        return TransactionType.WRAP
+      }
+      if (outputCurrency.isNative && WRAPPED_NATIVE_CURRENCY[chainId]?.equals(inputCurrency)) {
+        return TransactionType.UNWRAP
+      }
+    }
+    return undefined
+  }, [chainId, inputCurrency, outputCurrency])
+}
+
+export function useIsWrap(): boolean {
+  return useWrapType() !== undefined
+}
+
+export function useWrapCallback(
   wrapType?: TransactionType.WRAP | TransactionType.UNWRAP
 ): () => Promise<ContractTransaction | void> {
   const { account } = useWeb3React()
